requirements.txt
langchain==1.1.0
langchain_community==0.4.1
langchain_google_genai==3.2.0
langgraph==1.0.4
pydantic==2.12.5
python-dotenv==1.2.1
